Change Jolt physics structure for ragdoll development
