+++ WickedEngine/wiPhysics_Jolt.cpp
@@ -6,6 +6,7 @@
 #include "wiJobSystem.h"
 #include "wiRenderer.h"
 #include "wiTimer.h"
+#include "wiRagdoll.h"
 
 #include <Jolt/Jolt.h>
 #include <Jolt/RegisterTypes.h>
@@ -76,127 +77,29 @@ namespace wi::physics
 		bool ENABLED = true;
 		bool SIMULATION_ENABLED = true;
 		bool DEBUGDRAW_ENABLED = false;
-		int ACCURACY = 4;
 		int softbodyIterationCount = 6;
-		float TIMESTEP = 1.0f / 60.0f;
 		bool INTERPOLATION = true;
-
-		const uint cMaxBodies = 65536;
-		const uint cNumBodyMutexes = 0;
-		const uint cMaxBodyPairs = 65536;
-		const uint cMaxContactConstraints = 65536;
-		const EMotionQuality cMotionQuality = EMotionQuality::LinearCast;
-
-		inline Vec3 cast(const Float3& v) { return Vec3(v.x, v.y, v.z); }
-		inline Vec3 cast(const XMFLOAT3& v) { return Vec3(v.x, v.y, v.z); }
-		inline Quat cast(const XMFLOAT4& v) { return Quat(v.x, v.y, v.z, v.w); }
-		inline Mat44 cast(const XMFLOAT4X4& v)
+		
+		PhysicsScene& GetPhysicsScene(Scene& scene)
 		{
-			return Mat44(
-				Vec4(v._11, v._12, v._13, v._14),
-				Vec4(v._21, v._22, v._23, v._24),
-				Vec4(v._31, v._32, v._33, v._34),
-				Vec4(v._41, v._42, v._43, v._44)
-			);
+			if (scene.physics_scene == nullptr)
+			{
+				auto physics_scene = std::make_shared<PhysicsScene>();
+	
+				physics_scene->physics_system.Init(
+					cMaxBodies,
+					cNumBodyMutexes,
+					cMaxBodyPairs,
+					cMaxContactConstraints,
+					physics_scene->broad_phase_layer_interface,
+					physics_scene->object_vs_broadphase_layer_filter,
+					physics_scene->object_vs_object_layer_filter
+				);
+	
+				scene.physics_scene = physics_scene;
+			}
+			return *(PhysicsScene*)scene.physics_scene.get();
 		}
-		inline Mat44 cast(const XMMATRIX& v)
-		{
-			XMFLOAT4X4 m;
-			XMStoreFloat4x4(&m, v);
-			return cast(m);
-		}
-		inline XMFLOAT3 cast(Vec3Arg v) { return XMFLOAT3(v.GetX(), v.GetY(), v.GetZ()); }
-		inline XMFLOAT4 cast(QuatArg v) { return XMFLOAT4(v.GetX(), v.GetY(), v.GetZ(), v.GetW()); }
-		inline XMFLOAT4X4 cast(Mat44 v)
-		{
-			XMFLOAT4X4 ret;
-			v.StoreFloat4x4((Float4*)&ret);
-			return ret;
-		}
-
-		namespace Layers
-		{
-			static constexpr ObjectLayer NON_MOVING = 0;
-			static constexpr ObjectLayer MOVING = 1;
-			static constexpr ObjectLayer NUM_LAYERS = 2;
-		};
-
-		/// Class that determines if two object layers can collide
-		class ObjectLayerPairFilterImpl : public ObjectLayerPairFilter
-		{
-		public:
-			bool ShouldCollide(ObjectLayer inObject1, ObjectLayer inObject2) const override
-			{
-				switch (inObject1)
-				{
-				case Layers::NON_MOVING:
-					return inObject2 == Layers::MOVING; // Non moving only collides with moving
-				case Layers::MOVING:
-					return true; // Moving collides with everything
-				default:
-					JPH_ASSERT(false);
-					return false;
-				}
-			}
-		};
-
-		// Each broadphase layer results in a separate bounding volume tree in the broad phase. You at least want to have
-		// a layer for non-moving and moving objects to avoid having to update a tree full of static objects every frame.
-		// You can have a 1-on-1 mapping between object layers and broadphase layers (like in this case) but if you have
-		// many object layers you'll be creating many broad phase trees, which is not efficient. If you want to fine tune
-		// your broadphase layers define JPH_TRACK_BROADPHASE_STATS and look at the stats reported on the TTY.
-		namespace BroadPhaseLayers
-		{
-			static constexpr BroadPhaseLayer NON_MOVING(0);
-			static constexpr BroadPhaseLayer MOVING(1);
-			static constexpr uint NUM_LAYERS(2);
-		};
-
-		// BroadPhaseLayerInterface implementation
-		// This defines a mapping between object and broadphase layers.
-		class BPLayerInterfaceImpl final : public BroadPhaseLayerInterface
-		{
-		public:
-			BPLayerInterfaceImpl()
-			{
-				// Create a mapping table from object to broad phase layer
-				mObjectToBroadPhase[Layers::NON_MOVING] = BroadPhaseLayers::NON_MOVING;
-				mObjectToBroadPhase[Layers::MOVING] = BroadPhaseLayers::MOVING;
-			}
-
-			virtual uint GetNumBroadPhaseLayers() const override
-			{
-				return BroadPhaseLayers::NUM_LAYERS;
-			}
-
-			virtual BroadPhaseLayer GetBroadPhaseLayer(ObjectLayer inLayer) const override
-			{
-				JPH_ASSERT(inLayer < Layers::NUM_LAYERS);
-				return mObjectToBroadPhase[inLayer];
-			}
-
-		private:
-			BroadPhaseLayer mObjectToBroadPhase[Layers::NUM_LAYERS];
-		};
-
-		/// Class that determines if an object layer can collide with a broadphase layer
-		class ObjectVsBroadPhaseLayerFilterImpl : public ObjectVsBroadPhaseLayerFilter
-		{
-		public:
-			virtual bool ShouldCollide(ObjectLayer inLayer1, BroadPhaseLayer inLayer2) const override
-			{
-				switch (inLayer1)
-				{
-				case Layers::NON_MOVING:
-					return inLayer2 == BroadPhaseLayers::MOVING;
-				case Layers::MOVING:
-					return true;
-				default:
-					JPH_ASSERT(false);
-					return false;
-				}
-			}
-		};
 
 		struct JoltDestroyer
 		{
@@ -209,102 +112,6 @@ namespace wi::physics
 			}
 		} jolt_destroyer;
 
-		struct PhysicsScene
-		{
-			PhysicsSystem physics_system;
-			BPLayerInterfaceImpl broad_phase_layer_interface;
-			ObjectVsBroadPhaseLayerFilterImpl object_vs_broadphase_layer_filter;
-			ObjectLayerPairFilterImpl object_vs_object_layer_filter;
-			float accumulator = 0;
-			float alpha = 0;
-			bool activate_all_rigid_bodies = false;
-			float GetKinematicDT(float dt) const
-			{
-				return clamp(accumulator + dt, 0.0f, TIMESTEP * ACCURACY);
-			}
-		};
-		PhysicsScene& GetPhysicsScene(Scene& scene)
-		{
-			if (scene.physics_scene == nullptr)
-			{
-				auto physics_scene = std::make_shared<PhysicsScene>();
-
-				physics_scene->physics_system.Init(
-					cMaxBodies,
-					cNumBodyMutexes,
-					cMaxBodyPairs,
-					cMaxContactConstraints,
-					physics_scene->broad_phase_layer_interface,
-					physics_scene->object_vs_broadphase_layer_filter,
-					physics_scene->object_vs_object_layer_filter
-				);
-
-				scene.physics_scene = physics_scene;
-			}
-			return *(PhysicsScene*)scene.physics_scene.get();
-		}
-
-		struct RigidBody
-		{
-			std::shared_ptr<void> physics_scene;
-			ShapeRefC shape;
-			BodyID bodyID;
-			Entity entity = INVALID_ENTITY;
-			Vec3 initial_position = Vec3::sZero();
-			Quat initial_rotation = Quat::sIdentity();
-
-			// Things for parented objects:
-			XMFLOAT4X4 parentMatrix = wi::math::IDENTITY_MATRIX;
-			XMFLOAT4X4 parentMatrixInverse = wi::math::IDENTITY_MATRIX;
-
-			// Interpolation state:
-			Vec3 prev_position = Vec3::sZero();
-			Quat prev_rotation = Quat::sIdentity();
-
-			// Local body offset:
-			Mat44 additionalTransform = Mat44::sIdentity();
-			Mat44 additionalTransformInverse = Mat44::sIdentity();
-
-			// This is to fixup ragdolls that are applied to skeletons from different kinds of model imports
-			Mat44 restBasis = Mat44::sIdentity();
-			Mat44 restBasisInverse = Mat44::sIdentity();
-
-			// for trace hit reporting:
-			wi::ecs::Entity humanoid_ragdoll_entity = wi::ecs::INVALID_ENTITY;
-			wi::scene::HumanoidComponent::HumanoidBone humanoid_bone = wi::scene::HumanoidComponent::HumanoidBone::Count;
-			wi::primitive::Capsule capsule;
-
-			// water ripple stuff:
-			bool was_underwater = false;
-
-			// vehicle:
-			VehicleConstraint* vehicle_constraint = nullptr;
-			Vec3 prev_wheel_positions[4] = { Vec3::sZero(), Vec3::sZero(), Vec3::sZero(), Vec3::sZero() };
-			Quat prev_wheel_rotations[4] = { Quat::sIdentity(), Quat::sIdentity(), Quat::sIdentity(), Quat::sIdentity() };
-
-			void Delete()
-			{
-				if (physics_scene == nullptr || bodyID.IsInvalid())
-					return;
-				PhysicsScene* jolt_physics_scene = (PhysicsScene*)physics_scene.get();
-				BodyInterface& body_interface = jolt_physics_scene->physics_system.GetBodyInterface(); // locking version because destructor can be called from any thread
-				body_interface.RemoveBody(bodyID);
-				body_interface.DestroyBody(bodyID);
-				bodyID = {};
-				if (vehicle_constraint != nullptr)
-				{
-					jolt_physics_scene->physics_system.RemoveStepListener(vehicle_constraint);
-					jolt_physics_scene->physics_system.RemoveConstraint(vehicle_constraint);
-					vehicle_constraint = nullptr;
-				}
-				shape = nullptr;
-			}
-
-			~RigidBody()
-			{
-				Delete();
-			}
-		};
 		struct SoftBody
 		{
 			std::shared_ptr<void> physics_scene;
@@ -854,568 +661,7 @@ namespace wi::physics
 			}
 		}
 
-		struct Ragdoll
-		{
-			enum BODYPART
-			{
-				BODYPART_PELVIS = 0,
-				BODYPART_SPINE,
-				BODYPART_HEAD,
-
-				BODYPART_LEFT_UPPER_LEG,
-				BODYPART_LEFT_LOWER_LEG,
-
-				BODYPART_RIGHT_UPPER_LEG,
-				BODYPART_RIGHT_LOWER_LEG,
-
-				BODYPART_LEFT_UPPER_ARM,
-				BODYPART_LEFT_LOWER_ARM,
-
-				BODYPART_RIGHT_UPPER_ARM,
-				BODYPART_RIGHT_LOWER_ARM,
-
-				BODYPART_COUNT
-			};
-
-			std::shared_ptr<void> physics_scene;
-			RigidBody rigidbodies[BODYPART_COUNT];
-			Entity saved_parents[BODYPART_COUNT] = {};
-			Skeleton skeleton;
-			RagdollSettings settings;
-			Ref<JPH::Ragdoll> ragdoll;
-			bool state_active = false;
-			float scale = 1;
-			Vec3 prev_capsule_position[BODYPART_COUNT];
-			Quat prev_capsule_rotation[BODYPART_COUNT];
-
-			Ragdoll(Scene& scene, HumanoidComponent& humanoid, Entity humanoidEntity, float scale)
-			{
-				physics_scene = scene.physics_scene;
-				PhysicsSystem& physics_system = ((PhysicsScene*)physics_scene.get())->physics_system;
-				BodyInterface& body_interface = physics_system.GetBodyInterface(); // locking version because this is called from job system!
-
-				float masses[BODYPART_COUNT] = {};
-				Vec3 positions[BODYPART_COUNT] = {};
-				Vec3 constraint_positions[BODYPART_COUNT] = {};
-				Mat44 final_transforms[BODYPART_COUNT] = {};
-#if 0
-				// slow speed and visualizer to aid debugging:
-				wi::renderer::SetGameSpeed(0.1f);
-				SetDebugDrawEnabled(true);
-#endif
-
-				// Detect which way humanoid is facing in rest pose:
-				const float facing = humanoid.default_facing;
-
-				// Whole ragdoll will take a uniform scaling:
-				const XMMATRIX scaleMatrix = XMMatrixScaling(scale, scale, scale);
-				this->scale = scale;
-
-				// Calculate the bone lengths and radiuses in armature local space and create rigid bodies for bones:
-				for (int c = 0; c < BODYPART_COUNT; ++c)
-				{
-					HumanoidComponent::HumanoidBone humanoid_bone = HumanoidComponent::HumanoidBone::Count;
-					Entity entityA = INVALID_ENTITY;
-					Entity entityB = INVALID_ENTITY;
-					switch (c)
-					{
-					case BODYPART_PELVIS:
-						humanoid_bone = HumanoidComponent::HumanoidBone::Hips;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Hips];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Spine];
-						break;
-					case BODYPART_SPINE:
-						humanoid_bone = HumanoidComponent::HumanoidBone::Spine;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Spine];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Neck]; // prefer neck instead of head
-						if (entityB == INVALID_ENTITY)
-						{
-							entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Head]; // fall back to head if neck not available
-						}
-						break;
-					case BODYPART_HEAD:
-						humanoid_bone = HumanoidComponent::HumanoidBone::Neck;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Neck]; // prefer neck instead of head
-						if (entityA == INVALID_ENTITY)
-						{
-							humanoid_bone = HumanoidComponent::HumanoidBone::Head;
-							entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::Head]; // fall back to head if neck not available
-						}
-						break;
-					case BODYPART_LEFT_UPPER_LEG:
-						humanoid_bone = HumanoidComponent::HumanoidBone::LeftUpperLeg;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftUpperLeg];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftLowerLeg];
-						break;
-					case BODYPART_LEFT_LOWER_LEG:
-						humanoid_bone = HumanoidComponent::HumanoidBone::LeftLowerLeg;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftLowerLeg];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftFoot];
-						break;
-					case BODYPART_RIGHT_UPPER_LEG:
-						humanoid_bone = HumanoidComponent::HumanoidBone::RightUpperLeg;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightUpperLeg];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightLowerLeg];
-						break;
-					case BODYPART_RIGHT_LOWER_LEG:
-						humanoid_bone = HumanoidComponent::HumanoidBone::RightLowerLeg;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightLowerLeg];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightFoot];
-						break;
-					case BODYPART_LEFT_UPPER_ARM:
-						humanoid_bone = HumanoidComponent::HumanoidBone::LeftUpperArm;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftUpperArm];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftLowerArm];
-						break;
-					case BODYPART_LEFT_LOWER_ARM:
-						humanoid_bone = HumanoidComponent::HumanoidBone::LeftLowerArm;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftLowerArm];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::LeftHand];
-						break;
-					case BODYPART_RIGHT_UPPER_ARM:
-						humanoid_bone = HumanoidComponent::HumanoidBone::RightUpperArm;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightUpperArm];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightLowerArm];
-						break;
-					case BODYPART_RIGHT_LOWER_ARM:
-						humanoid_bone = HumanoidComponent::HumanoidBone::RightLowerArm;
-						entityA = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightLowerArm];
-						entityB = humanoid.bones[(size_t)HumanoidComponent::HumanoidBone::RightHand];
-						break;
-					}
-					if (entityA == INVALID_ENTITY)
-					{
-						wilog_warning("Ragdoll creation aborted because of a missing body part: %d", c);
-						return;
-					}
-
-					// Calculations here will be done in armature local space.
-					//	Unfortunately since humanoid can be separate from armature, we use a "find" utility to find bone rest matrix in armature
-					//	Note that current scaling of character is applied here separately from rest pose
-					XMMATRIX restA = scene.GetRestPose(entityA) * scaleMatrix;
-					XMMATRIX restB = scene.GetRestPose(entityB) * scaleMatrix;
-					XMVECTOR rootA = restA.r[3];
-					XMVECTOR rootB = restB.r[3];
-
-					// Every bone will be a rigid body:
-					RigidBody& physicsobject = rigidbodies[c];
-					physicsobject.entity = entityA;
-
-					float mass = scale;
-					float capsule_height = scale;
-					float capsule_radius = scale * humanoid.ragdoll_fatness;
-
-					if (c == BODYPART_HEAD)
-					{
-						// Head doesn't necessarily have a child, so make up something reasonable:
-						capsule_height = 0.05f * scale;
-						capsule_radius = 0.1f * scale * humanoid.ragdoll_headsize;
-					}
-					else
-					{
-						// bone length:
-						XMVECTOR len = XMVector3Length(XMVectorSubtract(rootB, rootA));
-						capsule_height = XMVectorGetX(len);
-
-						// capsule radius and length is tweaked per body part:
-						switch (c)
-						{
-						case BODYPART_PELVIS:
-							capsule_radius = 0.1f * scale * humanoid.ragdoll_fatness;
-							break;
-						case BODYPART_SPINE:
-							capsule_radius = 0.1f * scale * humanoid.ragdoll_fatness;
-							capsule_height -= capsule_radius * 2;
-							break;
-						case BODYPART_LEFT_LOWER_ARM:
-						case BODYPART_RIGHT_LOWER_ARM:
-							capsule_radius = capsule_height * 0.15f * humanoid.ragdoll_fatness;
-							capsule_height += capsule_radius;
-							break;
-						case BODYPART_LEFT_UPPER_LEG:
-						case BODYPART_RIGHT_UPPER_LEG:
-							capsule_radius = capsule_height * 0.15f * humanoid.ragdoll_fatness;
-							capsule_height -= capsule_radius * 2;
-							break;
-						case BODYPART_LEFT_LOWER_LEG:
-						case BODYPART_RIGHT_LOWER_LEG:
-							capsule_radius = capsule_height * 0.15f * humanoid.ragdoll_fatness;
-							capsule_height -= capsule_radius;
-							break;
-						default:
-							capsule_radius = capsule_height * 0.2f * humanoid.ragdoll_fatness;
-							capsule_height -= capsule_radius * 2;
-							break;
-						}
-					}
-
-					capsule_radius = std::abs(capsule_radius);
-					capsule_height = std::abs(capsule_height);
-
-					ShapeSettings::ShapeResult shape_result;
-					CapsuleShapeSettings shape_settings(capsule_height * 0.5f, capsule_radius);
-					shape_settings.SetEmbedded();
-					shape_result = shape_settings.Create();
-
-					RotatedTranslatedShapeSettings rtshape_settings;
-					rtshape_settings.SetEmbedded();
-					rtshape_settings.mInnerShapePtr = shape_result.Get();
-					rtshape_settings.mPosition = Vec3::sZero();
-					rtshape_settings.mRotation = Quat::sIdentity();
-
-					switch (c)
-					{
-					case BODYPART_LEFT_UPPER_ARM:
-					case BODYPART_LEFT_LOWER_ARM:
-					case BODYPART_RIGHT_UPPER_ARM:
-					case BODYPART_RIGHT_LOWER_ARM:
-						physicsobject.capsule = wi::primitive::Capsule(XMFLOAT3(-capsule_height * 0.5f - capsule_radius, 0, 0), XMFLOAT3(capsule_height * 0.5f + capsule_radius, 0, 0), capsule_radius);
-						rtshape_settings.mRotation = Quat::sRotation(Vec3::sAxisZ(), 0.5f * JPH_PI).Normalized();
-						break;
-					default:
-						physicsobject.capsule = wi::primitive::Capsule(XMFLOAT3(0, -capsule_height * 0.5f - capsule_radius, 0), XMFLOAT3(0, capsule_height * 0.5f + capsule_radius, 0), capsule_radius);
-						break;
-					}
-
-					shape_result = rtshape_settings.Create();
-					physicsobject.shape = shape_result.Get();
-
-					// capsule offset on axis is performed because otherwise capsule center would be in the bone root position
-					//	which is not what we want. Instead the bone is moved on its axis so it resides between root and tail
-					const float offset = capsule_height * 0.5f + capsule_radius;
-
-					Vec3 local_offset = Vec3::sZero();
-					switch (c)
-					{
-					case BODYPART_PELVIS:
-						break;
-					case BODYPART_SPINE:
-					case BODYPART_HEAD:
-						local_offset = Vec3(0, offset, 0);
-						break;
-					case BODYPART_LEFT_UPPER_LEG:
-					case BODYPART_LEFT_LOWER_LEG:
-					case BODYPART_RIGHT_UPPER_LEG:
-					case BODYPART_RIGHT_LOWER_LEG:
-						local_offset = Vec3(0, -offset, 0);
-						break;
-					case BODYPART_LEFT_UPPER_ARM:
-					case BODYPART_LEFT_LOWER_ARM:
-						local_offset = Vec3(-offset * facing, 0, 0);
-						break;
-					case BODYPART_RIGHT_UPPER_ARM:
-					case BODYPART_RIGHT_LOWER_ARM:
-						local_offset = Vec3(offset * facing, 0, 0);
-						break;
-					default:
-						break;
-					}
-
-					physicsobject.additionalTransform.SetTranslation(local_offset);
-					physicsobject.additionalTransformInverse = physicsobject.additionalTransform.Inversed();
-
-					// Get the translation and rotation part of rest matrix:
-					XMVECTOR SCA = {};
-					XMVECTOR ROT = {};
-					XMVECTOR TRA = {};
-					XMMatrixDecompose(&SCA, &ROT, &TRA, restA);
-					XMFLOAT4 rot = {};
-					XMFLOAT3 tra = {};
-					XMStoreFloat4(&rot, ROT);
-					XMStoreFloat3(&tra, TRA);
-
-					Vec3 root = cast(tra);
-
-					Mat44 mat = Mat44::sTranslation(root);
-					mat = mat * physicsobject.additionalTransform;
-
-					physicsobject.restBasis = Mat44::sRotation(cast(rot));
-					physicsobject.restBasisInverse = physicsobject.restBasis.Inversed();
-
-					physicsobject.humanoid_ragdoll_entity = humanoidEntity;
-					physicsobject.humanoid_bone = humanoid_bone;
-
-					physicsobject.physics_scene = scene.physics_scene;
-
-					masses[c] = mass;
-					positions[c] = mat.GetTranslation();
-					constraint_positions[c] = root;
-					final_transforms[c] = Mat44::sTranslation(cast(tra)) * Mat44::sRotation(cast(rot));
-					physicsobject.prev_position = final_transforms[c].GetTranslation();
-					physicsobject.prev_rotation = final_transforms[c].GetQuaternion().Normalized();
-					physicsobject.initial_position = physicsobject.prev_position;
-					physicsobject.initial_rotation = physicsobject.prev_rotation;
-					final_transforms[c] = final_transforms[c] * physicsobject.restBasisInverse;
-					final_transforms[c] = final_transforms[c] * physicsobject.additionalTransform;
-				}
-
-				// For constraint setup, see examples in Jolt/Samples/Utils/RagdollLoader.cpp
-
-				skeleton.SetEmbedded();
-
-				uint bodyparts[BODYPART_COUNT] = {};
-				bodyparts[BODYPART_PELVIS] = skeleton.AddJoint("LowerBody");
-				bodyparts[BODYPART_SPINE] = skeleton.AddJoint("UpperBody", bodyparts[BODYPART_PELVIS]);
-				bodyparts[BODYPART_HEAD] = skeleton.AddJoint("Head", bodyparts[BODYPART_SPINE]);
-				bodyparts[BODYPART_LEFT_UPPER_LEG] = skeleton.AddJoint("UpperLegL", bodyparts[BODYPART_PELVIS]);
-				bodyparts[BODYPART_LEFT_LOWER_LEG] = skeleton.AddJoint("LowerLegL", bodyparts[BODYPART_LEFT_UPPER_LEG]);
-				bodyparts[BODYPART_RIGHT_UPPER_LEG] = skeleton.AddJoint("UpperLegR", bodyparts[BODYPART_PELVIS]);
-				bodyparts[BODYPART_RIGHT_LOWER_LEG] = skeleton.AddJoint("LowerLegR", bodyparts[BODYPART_RIGHT_UPPER_LEG]);
-				bodyparts[BODYPART_LEFT_UPPER_ARM] = skeleton.AddJoint("UpperArmL", bodyparts[BODYPART_SPINE]);
-				bodyparts[BODYPART_LEFT_LOWER_ARM] = skeleton.AddJoint("LowerArmL", bodyparts[BODYPART_LEFT_UPPER_ARM]);
-				bodyparts[BODYPART_RIGHT_UPPER_ARM] = skeleton.AddJoint("UpperArmR", bodyparts[BODYPART_SPINE]);
-				bodyparts[BODYPART_RIGHT_LOWER_ARM] = skeleton.AddJoint("LowerArmR", bodyparts[BODYPART_RIGHT_UPPER_ARM]);
-
-				// Constraint limits
-				const float twist_angle[] = {
-					0.0f,		// Lower Body (unused, there's no parent)
-					5.0f,		// Upper Body
-					90.0f,		// Head
-					45.0f,		// Upper Leg L
-					45.0f,		// Lower Leg L
-					45.0f,		// Upper Leg R
-					45.0f,		// Lower Leg R
-					45.0f,		// Upper Arm L
-					45.0f,		// Lower Arm L
-					45.0f,		// Upper Arm R
-					45.0f,		// Lower Arm R
-				};
-
-				const float normal_angle[] = {
-					0.0f,		// Lower Body (unused, there's no parent)
-					40.0f,		// Upper Body
-					45.0f,		// Head
-					45.0f,		// Upper Leg L
-					0.0f,		// Lower Leg L
-					45.0f,		// Upper Leg R
-					0.0f,		// Lower Leg R
-					90.0f,		// Upper Arm L
-					0.0f,		// Lower Arm L
-					90.0f,		// Upper Arm R
-					0.0f,		// Lower Arm R
-				};
-
-				const float plane_angle[] = {
-					0.0f,		// Lower Body (unused, there's no parent)
-					40.0f,		// Upper Body
-					45.0f,		// Head
-					45.0f,		// Upper Leg L
-					60.0f,		// Lower Leg L (cheating here, a knee is not symmetric, we should have rotated the twist axis)
-					45.0f,		// Upper Leg R
-					60.0f,		// Lower Leg R
-					45.0f,		// Upper Arm L
-					90.0f,		// Lower Arm L
-					45.0f,		// Upper Arm R
-					90.0f,		// Lower Arm R
-				};
-
-				static float constraint_dbg = 0.1f;
-				static bool fixpose = false; // enable to fix the pose to rest pose, useful for debugging
-
-				settings.SetEmbedded();
-				settings.mSkeleton = &skeleton;
-				settings.mParts.resize(skeleton.GetJointCount());
-				for (int p = 0; p < skeleton.GetJointCount(); ++p)
-				{
-					RagdollSettings::Part& part = settings.mParts[p];
-					part.SetShape(rigidbodies[p].shape);
-					part.mPosition = positions[p];
-					part.mRotation = Quat::sIdentity();
-					part.mMotionType = EMotionType::Kinematic;
-					part.mMotionQuality = cMotionQuality;
-					part.mObjectLayer = Layers::MOVING;
-					part.mOverrideMassProperties = EOverrideMassProperties::CalculateInertia;
-					part.mMassPropertiesOverride.mMass = masses[p];
-					
-					// First part is the root, doesn't have a parent and doesn't have a constraint
-					if (p > 0)
-					{
-						if (p == BODYPART_LEFT_LOWER_LEG || p == BODYPART_RIGHT_LOWER_LEG)
-						{
-							Ref<HingeConstraintSettings> constraint = new HingeConstraintSettings;
-							constraint->mDrawConstraintSize = constraint_dbg;
-							constraint->mPoint1 = constraint->mPoint2 = constraint_positions[p];
-							constraint->mHingeAxis1 = constraint->mHingeAxis2 = Vec3::sAxisX() * facing;
-							constraint->mNormalAxis1 = constraint->mNormalAxis2 = -Vec3::sAxisY();
-							if (fixpose)
-							{
-								constraint->mLimitsMin = constraint->mLimitsMax = 0;
-							}
-							else
-							{
-								constraint->mLimitsMin = 0;
-								constraint->mLimitsMax = JPH_PI * 0.8f;
-							}
-							part.mToParent = constraint;
-						}
-						else if (p == BODYPART_LEFT_LOWER_ARM)
-						{
-							Ref<HingeConstraintSettings> constraint = new HingeConstraintSettings;
-							constraint->mDrawConstraintSize = constraint_dbg;
-							constraint->mPoint1 = constraint->mPoint2 = constraint_positions[p];
-							constraint->mHingeAxis1 = constraint->mHingeAxis2 = Vec3::sAxisY();
-							constraint->mNormalAxis1 = constraint->mNormalAxis2 = (constraint_positions[p] - constraint_positions[p - 1]).Normalized();
-							if (fixpose)
-							{
-								constraint->mLimitsMin = constraint->mLimitsMax = 0;
-							}
-							else
-							{
-								constraint->mLimitsMin = 0;
-								constraint->mLimitsMax = JPH_PI * 0.6f;
-							}
-							part.mToParent = constraint;
-						}
-						else if (p == BODYPART_RIGHT_LOWER_ARM)
-						{
-							Ref<HingeConstraintSettings> constraint = new HingeConstraintSettings;
-							constraint->mDrawConstraintSize = constraint_dbg;
-							constraint->mPoint1 = constraint->mPoint2 = constraint_positions[p];
-							constraint->mHingeAxis1 = constraint->mHingeAxis2 = -Vec3::sAxisY();
-							constraint->mNormalAxis1 = constraint->mNormalAxis2 = (constraint_positions[p] - constraint_positions[p - 1]).Normalized();
-							if (fixpose)
-							{
-								constraint->mLimitsMin = constraint->mLimitsMax = 0;
-							}
-							else
-							{
-								constraint->mLimitsMin = 0;
-								constraint->mLimitsMax = JPH_PI * 0.6f;
-							}
-							part.mToParent = constraint;
-						}
-						else
-						{
-							Ref<SwingTwistConstraintSettings> constraint = new SwingTwistConstraintSettings;
-							constraint->mDrawConstraintSize = constraint_dbg;
-							constraint->mPosition1 = constraint->mPosition2 = constraint_positions[p];
-							constraint->mTwistAxis1 = constraint->mTwistAxis2 = (positions[p] - constraint_positions[p]).Normalized();
-							constraint->mPlaneAxis1 = constraint->mPlaneAxis2 = Vec3::sAxisZ() * facing;
-							if (fixpose)
-							{
-								constraint->mTwistMinAngle = constraint->mTwistMaxAngle = 0;
-								constraint->mNormalHalfConeAngle = 0;
-								constraint->mPlaneHalfConeAngle = 0;
-							}
-							else
-							{
-								constraint->mTwistMinAngle = -DegreesToRadians(twist_angle[p]);
-								constraint->mTwistMaxAngle = DegreesToRadians(twist_angle[p]);
-								constraint->mNormalHalfConeAngle = DegreesToRadians(normal_angle[p]);
-								constraint->mPlaneHalfConeAngle = DegreesToRadians(plane_angle[p]);
-							}
-							part.mToParent = constraint;
-						}
-					}
-				}
-
-				settings.Stabilize();
-				settings.DisableParentChildCollisions();
-				settings.CalculateBodyIndexToConstraintIndex();
-
-				static std::atomic<uint32_t> collisionGroupID{}; // generate unique collision group for each ragdoll to enable collision between them
-				ragdoll = settings.CreateRagdoll(collisionGroupID.fetch_add(1), 0, &physics_system);
-				ragdoll->SetPose(Vec3::sZero(), final_transforms);
-				ragdoll->AddToPhysicsSystem(EActivation::Activate);
-
-				const int count = (int)ragdoll->GetBodyCount();
-				for (int index = 0; index < count; ++index)
-				{
-					rigidbodies[index].bodyID = ragdoll->GetBodyID(index);
-					body_interface.SetUserData(rigidbodies[index].bodyID, (uint64_t)&rigidbodies[index]);
-				}
-			}
-			~Ragdoll()
-			{
-				if (physics_scene == nullptr)
-					return;
-				PhysicsSystem& physics_system = ((PhysicsScene*)physics_scene.get())->physics_system;
-
-				const int count = (int)ragdoll->GetBodyCount();
-				for (int index = 0; index < count; ++index)
-				{
-					rigidbodies[index].bodyID = {}; // will be removed by ragdoll
-				}
-
-				ragdoll->RemoveFromPhysicsSystem();
-			}
-
-			// Activates ragdoll as dynamic physics object:
-			void Activate(
-				Scene& scene,
-				Entity humanoidEntity
-			)
-			{
-				if (state_active)
-					return;
-				state_active = true;
-
-				const HumanoidComponent* humanoid = scene.humanoids.GetComponent(humanoidEntity);
-				if (humanoid == nullptr)
-					return;
-
-				PhysicsSystem& physics_system = ((PhysicsScene*)physics_scene.get())->physics_system;
-				BodyInterface& body_interface = physics_system.GetBodyInterface(); // locking version because this is called from job system!
-
-				int c = 0;
-				for (auto& x : rigidbodies)
-				{
-					body_interface.SetMotionType(x.bodyID, EMotionType::Dynamic, EActivation::Activate);
-
-					// Save parenting information to be able to restore it:
-					const HierarchyComponent* hier = scene.hierarchy.GetComponent(x.entity);
-					if (hier != nullptr)
-					{
-						saved_parents[c] = hier->parentID;
-					}
-					else
-					{
-						saved_parents[c] = INVALID_ENTITY;
-					}
-
-					// detach bone because it will be simulated in world space:
-					scene.Component_Detach(x.entity);
-
-					c++;
-				}
-
-				// Stop all anims that are children of humanoid:
-				for (size_t i = 0; i < scene.animations.GetCount(); ++i)
-				{
-					Entity entity = scene.animations.GetEntity(i);
-					if (!scene.Entity_IsDescendant(entity, humanoidEntity))
-						continue;
-					AnimationComponent& animation = scene.animations[i];
-					animation.Stop();
-				}
-			}
-
-			// Disables dynamic ragdoll and reattaches loose parts as they were:
-			void Deactivate(
-				Scene& scene
-			)
-			{
-				if (!state_active)
-					return;
-				state_active = false;
-
-				PhysicsSystem& physics_system = ((PhysicsScene*)physics_scene.get())->physics_system;
-				BodyInterface& body_interface = physics_system.GetBodyInterface(); // locking version because this is called from job system!
-
-				int c = 0;
-				for (auto& x : rigidbodies)
-				{
-					body_interface.SetMotionType(x.bodyID, EMotionType::Kinematic, EActivation::Activate);
-
-					if (saved_parents[c] != INVALID_ENTITY)
-					{
-						scene.Component_Attach(x.entity, saved_parents[c]);
-					}
-					c++;
-				}
-			}
-		};
-
+		
 	}
 	using namespace jolt;
 
@@ -2129,9 +1375,9 @@ namespace wi::physics
 				return;
 			Ragdoll& ragdoll = *(Ragdoll*)humanoid.ragdoll.get();
 
-			if (humanoid.ragdoll_bodyparts.size() != Ragdoll::BODYPART_COUNT)
+			if (humanoid.ragdoll_bodyparts.size() != ragdoll.getSize())
 			{
-				humanoid.ragdoll_bodyparts.resize(Ragdoll::BODYPART_COUNT);
+				humanoid.ragdoll_bodyparts.resize(ragdoll.getSize());
 			}
 			humanoid.ragdoll_bounds = wi::primitive::AABB();
 			int bodypart = 0;
